Added an attack speed variable and rearranged conditionals
Now sets a timer each time a player attacks that prevents the player
from attacking again until it runs out. The length of the timer is based
on the public weaponSpeed variable, which should be changed to alter
attack speed. I defaulted it to 8 seconds for now. Still needs work. I
also rearranged the conditionals so as to allow the player to attack
even if there is no other player in the attack range. This makes it more
realistic, while also giving players more of a challenge to hit
eachother since a swing doesn't guarantee a hit anymore.